+++ ide/fplidemenu.py
@@ -1,6 +1,7 @@
 from tkinter import messagebox
 from ide.DialogSettings import DialogSettings
-from ide.DialogNewFplTheory import DialogNewFplTheory
+from ide.DialogNewOrAddFpl import DialogNewOrAddFpl
+from ide.DialogOpenFplTheory import DialogOpenFplTheory
 import tkinter as tk
 
 
@@ -12,10 +13,10 @@ class FPLIdeMenus:
 
         self.fpl_theory_menu = tk.Menu(self._menuBar, tearoff=0)
         self.fpl_theory_menu.add_command(label='New', underline=0, command=self.new_fpl_theory)
-        self.fpl_theory_menu.add_command(label='Open', underline=0, command=self.not_implemented)
-        self.fpl_theory_menu.add_command(label='Close', underline=0, command=self.not_implemented)
+        self.fpl_theory_menu.add_command(label='Open', underline=0, command=self.open_fpl_theory)
+        self.fpl_theory_menu.add_command(label='Close', underline=0, command=self.close_fpl_theory)
         self.fpl_theory_menu.add_separator()
-        self.fpl_theory_menu.add_command(label='Add FPL file', underline=0, command=self.not_implemented)
+        self.fpl_theory_menu.add_command(label='Add FPL file', underline=0, command=self.add_fpl_file)
         self.fpl_theory_menu.add_command(label='Remove FPL file', underline=0, command=self.not_implemented)
         self.fpl_theory_menu.add_separator()
         self.fpl_theory_menu.add_command(label='Exit', underline=1, command=self.exit)
@@ -90,22 +91,8 @@ class FPLIdeMenus:
             self.fpl_theory_menu.entryconfig("Remove FPL file", state=tk.DISABLED)
 
     def exit(self, event=None):
-        book = self.ide.get_editor_notebook().get_files()
-        at_least_one_open_file_changed = False
-        for file in book:
-            if book[file].text.is_dirty:
-                at_least_one_open_file_changed = True
-                self.ide.get_editor_notebook().select_file(file)
-                msg = messagebox.askyesnocancel("Quit FPLIDE",
-                                                "Do you want to save changes of the file " + file + " before quitting?",
-                                                icon='warning')
-                if msg:
-                    # the user does not want to reopen the file, just select the tab!
-                    self.save_file(None)
-                elif msg is None:
-                    # do nothing
-                    return
-        if not at_least_one_open_file_changed:
+        closeable = self._ask_to_save_any_open_files()
+        if closeable:
             # if no message boxes were answered yet, ask if the user really want's to quit the application
             msg = messagebox.askyesnocancel("Quit FPLIDE",
                                             "Do you want quit the application?",
@@ -122,4 +109,49 @@ class FPLIdeMenus:
         DialogSettings(self.ide)
 
     def new_fpl_theory(self):
-        DialogNewFplTheory(self.ide)
+        DialogNewOrAddFpl(self.ide, "new")
+
+    def open_fpl_theory(self):
+        DialogOpenFplTheory(self.ide)
+
+    def add_fpl_file(self):
+        DialogNewOrAddFpl(self.ide, "add")
+
+    def _ask_to_save_any_open_files(self):
+        book = self.ide.get_editor_notebook().get_files()
+        for file in book:
+            if book[file].text.is_dirty:
+                self.ide.get_editor_notebook().select_file(file)
+                msg = messagebox.askyesnocancel("Quit FPLIDE",
+                                                "Do you want to save changes of the file " + file + " before quitting?",
+                                                icon='warning')
+                if msg:
+                    # the user wants to save the file!
+                    self.save_file(None)
+                elif msg is None:
+                    # the user cancelled the exist
+                    return False
+        # the user decided if to save each file and saved it.
+        return True
+
+    def close_fpl_theory(self):
+        closeable = self._ask_to_save_any_open_files()
+        if closeable:
+            # close all open tabs
+            book = self.ide.get_editor_notebook()
+            for tab_name in book.tabs():
+                book.forget(tab_name)
+            # empty error list
+            error_tree_view = self.ide.get_error_list()
+            error_tree_view.delete(*error_tree_view.get_children())
+            self.ide.update_error_warning_counts()
+            # empty object tree
+            object_tree = self.ide.get_object_browser_tree()
+            object_tree.delete(*object_tree.get_children())
+            # flag that no theory is open
+            self.ide.model.theory_is_open_flag = False
+            self.ide.menus.menu_configure()
+            # set the status bar
+            status_bar = self.ide.get_status_bar()
+            status_bar.set_text("Ready.")
+